feat: add balance for the Graph Staking on Arbitrum One
